Create test_mcp_server_lifecycle.py with 27 lifecycle tests
Completed Task features.bees-xhi and review fix features.bees-3jb:
- Created test_mcp_server_lifecycle.py (407 lines) with server lifecycle tests
- Extracted 6 test classes: TestMCPServerInitialization, TestServerLifecycle,
  TestHealthCheck, TestToolRegistration, TestErrorHandling, TestServerConfiguration,
  and TestModuleIntegration
- Removed lifecycle tests from test_mcp_server.py (~2790 lines remaining)
- Updated README.md and master_plan.md with test organization documentation
- All 1316 tests pass
