+++ docs/tickets/features/features.bees-5y8.md
@@ -38,8 +38,9 @@ children:
 - features.bees-82b
 - features.bees-xab
 - features.bees-se5
+- features.bees-3jb
 created_at: '2026-02-05T08:05:49.511857'
-updated_at: '2026-02-05T16:15:53.853627'
+updated_at: '2026-02-05T16:30:29.830552'
 priority: 2
 status: open
 bees_version: '1.1'
