+++ docs/tickets/features/features.bees-m9r.md
@@ -1,0 +1,50 @@
+---
+id: features.bees-m9r
+type: subtask
+title: Run unit tests and fix failures
+description: 'Execute pytest test suite to verify TestModuleIntegration extraction
+  was successful.
+
+
+  What to do:
+
+  - Run `pytest tests/test_mcp_server_lifecycle.py -v` to verify extracted tests pass
+
+  - Run `pytest tests/test_mcp_server.py -v` to verify remaining tests pass
+
+  - Run full suite `pytest tests/test_mcp_*.py` to ensure no regressions
+
+  - Fix any failures until 100% pass
+
+
+  Success criteria:
+
+  - All tests in test_mcp_server_lifecycle.py pass
+
+  - All tests in test_mcp_server.py pass
+
+  - Total test count remains 155 tests
+
+  - No test failures or import errors'
+up_dependencies:
+- features.bees-k1p
+parent: features.bees-3jb
+created_at: '2026-02-05T16:31:03.502381'
+updated_at: '2026-02-05T16:32:58.006419'
+status: completed
+bees_version: '1.1'
+---
+
+Execute pytest test suite to verify TestModuleIntegration extraction was successful.
+
+What to do:
+- Run `pytest tests/test_mcp_server_lifecycle.py -v` to verify extracted tests pass
+- Run `pytest tests/test_mcp_server.py -v` to verify remaining tests pass
+- Run full suite `pytest tests/test_mcp_*.py` to ensure no regressions
+- Fix any failures until 100% pass
+
+Success criteria:
+- All tests in test_mcp_server_lifecycle.py pass
+- All tests in test_mcp_server.py pass
+- Total test count remains 155 tests
+- No test failures or import errors
