+++ docs/tickets/features/features.bees-q20.md
@@ -29,8 +29,8 @@ down_dependencies:
 - features.bees-rfd
 parent: features.bees-xhi
 created_at: '2026-02-05T16:14:03.169160'
-updated_at: '2026-02-05T16:14:08.148883'
-status: open
+updated_at: '2026-02-05T16:27:05.253236'
+status: completed
 bees_version: '1.1'
 ---
