+++ docs/tickets/features/features.bees-uxt.md
@@ -21,12 +21,12 @@ description: 'Context: Test suite has been reorganized into multiple focused fil
 
 
   Acceptance: README.md documents test file organization and how to run tests'
-parent: features.bees-xhi
 up_dependencies:
 - features.bees-vw7
-status: open
+parent: features.bees-xhi
 created_at: '2026-02-05T16:13:52.479243'
-updated_at: '2026-02-05T16:13:52.479254'
+updated_at: '2026-02-05T16:26:24.369427'
+status: completed
 bees_version: '1.1'
 ---
