+++ docs/tickets/features/features.bees-vii.md
@@ -21,12 +21,12 @@ description: 'Context: Test file reorganization is an architectural change that 
 
 
   Acceptance: master_plan.md documents test architecture and reorganization rationale'
-parent: features.bees-xhi
 up_dependencies:
 - features.bees-vw7
-status: open
+parent: features.bees-xhi
 created_at: '2026-02-05T16:13:58.578436'
-updated_at: '2026-02-05T16:13:58.578441'
+updated_at: '2026-02-05T16:27:01.241131'
+status: completed
 bees_version: '1.1'
 ---
