+++ docs/tickets/features/features.bees-vw7.md
@@ -27,8 +27,8 @@ down_dependencies:
 - features.bees-q20
 parent: features.bees-xhi
 created_at: '2026-02-05T16:13:41.285721'
-updated_at: '2026-02-05T16:14:03.175884'
-status: open
+updated_at: '2026-02-05T16:24:15.071514'
+status: completed
 bees_version: '1.1'
 ---
